+++ clients/derivativestradingcoinfutures/src/restapi/models/model_cancel_order_response.go
@@ -17,14 +17,10 @@ var _ common.MappedNullable = &CancelOrderResponse{}
 
 // CancelOrderResponse struct for CancelOrderResponse
 type CancelOrderResponse struct {
-	// Average execution price. Will be removed after CM migration.
-	AvgPrice *string `json:"avgPrice,omitempty"`
 	// Client order ID.
 	ClientOrderId *string `json:"clientOrderId,omitempty"`
 	// Cumulative filled quantity.
 	CumQty *string `json:"cumQty,omitempty"`
-	// Cumulative base asset amount. Will be removed after CM migration.
-	CumBase *string `json:"cumBase,omitempty"`
 	// Executed quantity
 	ExecutedQty *string `json:"executedQty,omitempty"`
 	// Sub-order ID
@@ -91,38 +87,6 @@ func NewCancelOrderResponseWithDefaults() *CancelOrderResponse {
 	return &this
 }
 
-// GetAvgPrice returns the AvgPrice field value if set, zero value otherwise.
-func (o *CancelOrderResponse) GetAvgPrice() string {
-	if o == nil || common.IsNil(o.AvgPrice) {
-		var ret string
-		return ret
-	}
-	return *o.AvgPrice
-}
-
-// GetAvgPriceOk returns a tuple with the AvgPrice field value if set, nil otherwise
-// and a boolean to check if the value has been set.
-func (o *CancelOrderResponse) GetAvgPriceOk() (*string, bool) {
-	if o == nil || common.IsNil(o.AvgPrice) {
-		return nil, false
-	}
-	return o.AvgPrice, true
-}
-
-// HasAvgPrice returns a boolean if a field has been set.
-func (o *CancelOrderResponse) HasAvgPrice() bool {
-	if o != nil && !common.IsNil(o.AvgPrice) {
-		return true
-	}
-
-	return false
-}
-
-// SetAvgPrice gets a reference to the given string and assigns it to the AvgPrice field.
-func (o *CancelOrderResponse) SetAvgPrice(v string) {
-	o.AvgPrice = &v
-}
-
 // GetClientOrderId returns the ClientOrderId field value if set, zero value otherwise.
 func (o *CancelOrderResponse) GetClientOrderId() string {
 	if o == nil || common.IsNil(o.ClientOrderId) {
@@ -185,38 +149,6 @@ func (o *CancelOrderResponse) HasCumQty() bool {
 // SetCumQty gets a reference to the given string and assigns it to the CumQty field.
 func (o *CancelOrderResponse) SetCumQty(v string) {
 	o.CumQty = &v
-}
-
-// GetCumBase returns the CumBase field value if set, zero value otherwise.
-func (o *CancelOrderResponse) GetCumBase() string {
-	if o == nil || common.IsNil(o.CumBase) {
-		var ret string
-		return ret
-	}
-	return *o.CumBase
-}
-
-// GetCumBaseOk returns a tuple with the CumBase field value if set, nil otherwise
-// and a boolean to check if the value has been set.
-func (o *CancelOrderResponse) GetCumBaseOk() (*string, bool) {
-	if o == nil || common.IsNil(o.CumBase) {
-		return nil, false
-	}
-	return o.CumBase, true
-}
-
-// HasCumBase returns a boolean if a field has been set.
-func (o *CancelOrderResponse) HasCumBase() bool {
-	if o != nil && !common.IsNil(o.CumBase) {
-		return true
-	}
-
-	return false
-}
-
-// SetCumBase gets a reference to the given string and assigns it to the CumBase field.
-func (o *CancelOrderResponse) SetCumBase(v string) {
-	o.CumBase = &v
 }
 
 // GetExecutedQty returns the ExecutedQty field value if set, zero value otherwise.
@@ -933,17 +865,11 @@ func (o CancelOrderResponse) MarshalJSON() ([]byte, error) {
 
 func (o CancelOrderResponse) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
-	if !common.IsNil(o.AvgPrice) {
-		toSerialize["avgPrice"] = o.AvgPrice
-	}
 	if !common.IsNil(o.ClientOrderId) {
 		toSerialize["clientOrderId"] = o.ClientOrderId
 	}
 	if !common.IsNil(o.CumQty) {
 		toSerialize["cumQty"] = o.CumQty
-	}
-	if !common.IsNil(o.CumBase) {
-		toSerialize["cumBase"] = o.CumBase
 	}
 	if !common.IsNil(o.ExecutedQty) {
 		toSerialize["executedQty"] = o.ExecutedQty
@@ -1033,10 +959,8 @@ func (o *CancelOrderResponse) UnmarshalJSON(data []byte) (err error) {
 	additionalProperties := make(map[string]interface{})
 
 	if err = json.Unmarshal(data, &additionalProperties); err == nil {
-		delete(additionalProperties, "avgPrice")
 		delete(additionalProperties, "clientOrderId")
 		delete(additionalProperties, "cumQty")
-		delete(additionalProperties, "cumBase")
 		delete(additionalProperties, "executedQty")
 		delete(additionalProperties, "orderId")
 		delete(additionalProperties, "origQty")