+++ clients/derivativestradingcoinfutures/src/restapi/models/model_modify_order_response.go
@@ -29,16 +29,12 @@ type ModifyOrderResponse struct {
 	ClientOrderId *string `json:"clientOrderId,omitempty"`
 	// Latest token price.
 	Price *string `json:"price,omitempty"`
-	// Average execution price. Will be removed after CM migration.
-	AvgPrice *string `json:"avgPrice,omitempty"`
 	// Original order quantity
 	OrigQty *string `json:"origQty,omitempty"`
 	// Executed quantity
 	ExecutedQty *string `json:"executedQty,omitempty"`
 	// Cumulative filled quantity.
 	CumQty *string `json:"cumQty,omitempty"`
-	// Cumulative base asset amount. Will be removed after CM migration.
-	CumBase *string `json:"cumBase,omitempty"`
 	// Time in force
 	TimeInForce *string `json:"timeInForce,omitempty"`
 	// Order type.
@@ -279,38 +275,6 @@ func (o *ModifyOrderResponse) SetPrice(v string) {
 	o.Price = &v
 }
 
-// GetAvgPrice returns the AvgPrice field value if set, zero value otherwise.
-func (o *ModifyOrderResponse) GetAvgPrice() string {
-	if o == nil || common.IsNil(o.AvgPrice) {
-		var ret string
-		return ret
-	}
-	return *o.AvgPrice
-}
-
-// GetAvgPriceOk returns a tuple with the AvgPrice field value if set, nil otherwise
-// and a boolean to check if the value has been set.
-func (o *ModifyOrderResponse) GetAvgPriceOk() (*string, bool) {
-	if o == nil || common.IsNil(o.AvgPrice) {
-		return nil, false
-	}
-	return o.AvgPrice, true
-}
-
-// HasAvgPrice returns a boolean if a field has been set.
-func (o *ModifyOrderResponse) HasAvgPrice() bool {
-	if o != nil && !common.IsNil(o.AvgPrice) {
-		return true
-	}
-
-	return false
-}
-
-// SetAvgPrice gets a reference to the given string and assigns it to the AvgPrice field.
-func (o *ModifyOrderResponse) SetAvgPrice(v string) {
-	o.AvgPrice = &v
-}
-
 // GetOrigQty returns the OrigQty field value if set, zero value otherwise.
 func (o *ModifyOrderResponse) GetOrigQty() string {
 	if o == nil || common.IsNil(o.OrigQty) {
@@ -405,38 +369,6 @@ func (o *ModifyOrderResponse) HasCumQty() bool {
 // SetCumQty gets a reference to the given string and assigns it to the CumQty field.
 func (o *ModifyOrderResponse) SetCumQty(v string) {
 	o.CumQty = &v
-}
-
-// GetCumBase returns the CumBase field value if set, zero value otherwise.
-func (o *ModifyOrderResponse) GetCumBase() string {
-	if o == nil || common.IsNil(o.CumBase) {
-		var ret string
-		return ret
-	}
-	return *o.CumBase
-}
-
-// GetCumBaseOk returns a tuple with the CumBase field value if set, nil otherwise
-// and a boolean to check if the value has been set.
-func (o *ModifyOrderResponse) GetCumBaseOk() (*string, bool) {
-	if o == nil || common.IsNil(o.CumBase) {
-		return nil, false
-	}
-	return o.CumBase, true
-}
-
-// HasCumBase returns a boolean if a field has been set.
-func (o *ModifyOrderResponse) HasCumBase() bool {
-	if o != nil && !common.IsNil(o.CumBase) {
-		return true
-	}
-
-	return false
-}
-
-// SetCumBase gets a reference to the given string and assigns it to the CumBase field.
-func (o *ModifyOrderResponse) SetCumBase(v string) {
-	o.CumBase = &v
 }
 
 // GetTimeInForce returns the TimeInForce field value if set, zero value otherwise.
@@ -883,9 +815,6 @@ func (o ModifyOrderResponse) ToMap() (map[string]interface{}, error) {
 	if !common.IsNil(o.Price) {
 		toSerialize["price"] = o.Price
 	}
-	if !common.IsNil(o.AvgPrice) {
-		toSerialize["avgPrice"] = o.AvgPrice
-	}
 	if !common.IsNil(o.OrigQty) {
 		toSerialize["origQty"] = o.OrigQty
 	}
@@ -894,9 +823,6 @@ func (o ModifyOrderResponse) ToMap() (map[string]interface{}, error) {
 	}
 	if !common.IsNil(o.CumQty) {
 		toSerialize["cumQty"] = o.CumQty
-	}
-	if !common.IsNil(o.CumBase) {
-		toSerialize["cumBase"] = o.CumBase
 	}
 	if !common.IsNil(o.TimeInForce) {
 		toSerialize["timeInForce"] = o.TimeInForce
@@ -965,11 +891,9 @@ func (o *ModifyOrderResponse) UnmarshalJSON(data []byte) (err error) {
 		delete(additionalProperties, "status")
 		delete(additionalProperties, "clientOrderId")
 		delete(additionalProperties, "price")
-		delete(additionalProperties, "avgPrice")
 		delete(additionalProperties, "origQty")
 		delete(additionalProperties, "executedQty")
 		delete(additionalProperties, "cumQty")
-		delete(additionalProperties, "cumBase")
 		delete(additionalProperties, "timeInForce")
 		delete(additionalProperties, "type")
 		delete(additionalProperties, "reduceOnly")